The current article parsing code uses Start - 1 as a marker for various things, and using an unsigned index just isn't going to work easily.  After reviewing the original patch that made this change, long should work as well as size_t, so use that for right now.  This is fixed properly in CURRENT.

+++ innd/innd.h
@@ -223,16 +223,16 @@ typedef struct _ARTDATA {
   struct buffer	  Overview;		/* buffer for overview data */
   int		  CRwithoutLF;		/* counter for '\r' without '\n' */
   int		  LFwithoutCR;		/* counter for '\n' without '\r' */
-  size_t	  CurHeader;		/* where current header starts.
+  long		  CurHeader;		/* where current header starts.
 					   this is used for folded header
 					   it indicates offset from bp->Data */
   bool		  NullHeader;		/* contains NULL in current header    */
-  size_t	  LastTerminator;	/* where last '.' exists.  only set if
+  long		  LastTerminator;	/* where last '.' exists.  only set if
 					   it exists at the begining of line
 					   it indicates offset from bp->Data */
-  size_t	  LastCR;		/* where last CR exists
+  long		  LastCR;		/* where last CR exists
 					   it indicates offset from bp->Data */
-  size_t	  LastCRLF;		/* where last CRLF exists.
+  long		  LastCRLF;		/* where last CRLF exists.
 					   indicates where last LF exists
 					   it indicates offset from bp->Data */
   HDRCONTENT	  HdrContent[MAX_ARTHEADER];
